A-Packages: separate classes into duke and duke.command
