+++ src/main/java/duke/Deadline.java
@@ -1,6 +1,7 @@
+package duke;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-
 
 public class Deadline extends Task {
     private LocalDate by;
